Move connect() back to index.js
