+++ lib/sync/index.js
@@ -3,6 +3,8 @@ var metricsModule = require('./sync-metrics');
 var _ = require('underscore');
 var server = require('./sync-server');
 var util = require('util');
+var MongoClient = require('mongodb').MongoClient;
+var redis = require('redis');
 
 function toJSON(dataset_id, returnData, cb) {
   syncUtil.doLog(dataset_id, 'debug', 'toJSON');
@@ -22,6 +24,45 @@ function listCollisions(dataset_id, params, cb) {
 function removeCollision(dataset_id, params, cb) {
   syncUtil.doLog(dataset_id, 'info', 'removeCollision');
   return server.api.removeCollision(dataset_id, params.hash, params.meta_data, cb);
+}
+
+/**
+ * Connect to mongodb & redis with the given connection urls.
+ * This should be called before `start()` is called
+ *
+ * @param {string} mongoDBConnectionUrl A MongoDB connection uril in a format supported by the `mongodb` module
+ * @param {string} redisUrl A redis connection url in a format supported by the `redis` module
+ * @param {Object} metricsConf Metrics configuration
+ * @param {function} cb
+ */
+function connect(mongoDBConnectionUrl, redisUrl, metricsConf, cb) {
+  if (arguments.length < 4) throw new Error('connect requires 4 arguments');
+
+  async.series([
+    function connectToMongoDB(callback) {
+      MongoClient.connect(mongoDBConnectionUrl, callback);
+    },
+    function connectToRedis(callback) {
+      if (redisUrl) {
+        var redisOpts = {
+          url: redisUrl
+        };
+        var client = redis.createClient(redisOpts);
+        return callback(null, client);
+      } else {
+        return callback();
+      }
+    }
+  ], function(err, results) {
+    if (err) return cb(err);
+    var mongoDbClient = results[0];
+    var redisClient = results[1];
+    var metricsClient = metricsModule.init(metricsConf);
+
+    server.setClients(mongoDbClient, redisClient, metricsClient);
+
+    return cb(null, mongoDbClient, redisClient, metricsClient);
+  });
 }
 
 function setLogLevel(dataset_id, params, cb) {
@@ -69,6 +110,7 @@ function invoke(dataset_id, params, callback) {
 
 // extend from empty object to force copy, and because last source overrides the previous ones
 module.exports = _.extend({}, server, {
+  connect: connect,
   sync: metricsModule.timeAsyncFunc(metricsModule.KEYS.SYNC_API_PROCESS_TIME, server.sync),
   syncRecords: metricsModule.timeAsyncFunc(metricsModule.KEYS.SYNC_API_PROCESS_TIME, server.syncRecords),
   listCollisions: metricsModule.timeAsyncFunc(metricsModule.KEYS.SYNC_API_PROCESS_TIME, listCollisions),
