+++ lib/sync/sync-server.js
@@ -5,11 +5,8 @@ var datasets = require('./datasets');
 var defaultDataHandlersModule = require('./default-dataHandlers');
 var hashProviderModule = require('./hashProvider');
 var interceptorsModule = require('./interceptors');
-var metricsModule = require('./sync-metrics');
-var MongoClient = require('mongodb').MongoClient;
 var MongodbQueue = require('./mongodbQueue');
 var pendingProcessor = require('./pending-processor');
-var redis = require('redis');
 var storageModule = require('./storage');
 var syncApiModule = require('./api-sync');
 var syncLockModule = require('./lock');
@@ -67,50 +64,19 @@ function init(dataset_id, options, cb) {
   });
 }
 
-/**
- * Connect to mongodb & redis with the given connection urls.
- * This should be called before `start()` is called
- *
- * @param {string} mongoDBConnectionUrl A MongoDB connection uril in a format supported by the `mongodb` module
- * @param {string} redisUrl A redis connection url in a format supported by the `redis` module
- * @param {Object} metricsConf Metrics configuration
- * @param {function} cb
- */
-function connect(mongoDBConnectionUrl, redisUrl, metricsConf, cb) {
-  if (arguments.length < 4) throw new Error('connect requires 4 arguments');
-
-  async.series([
-    function connectToMongoDB(callback) {
-      MongoClient.connect(mongoDBConnectionUrl, callback);
-    },
-    function connectToRedis(callback) {
-      if (redisUrl) {
-        var redisOpts = {
-          url: redisUrl
-        };
-        var client = redis.createClient(redisOpts);
-        return callback(null, client);
-      } else {
-        return callback();
-      }
-    }
-  ], function(err, results) {
-    if (err) return cb(err);
-    mongoDbClient = results[0];
-    redisClient = results[1];
-    metricsClient = metricsModule.init(metricsConf);
-    dataHandlers = dataHandlersModule({
-      defaultHandlers: defaultDataHandlersModule(mongoDbClient)
-    });
-    syncStorage = storageModule(mongoDbClient);
-    // TODO: follow same pattern as other modules for this hashProviderModule
-    hashProvider = hashProviderModule;
-    syncLock = syncLockModule(mongoDbClient, 'fhsync_locks');
-    interceptors = interceptorsModule();
-    return cb(null, mongoDbClient, redisClient, metricsClient);
+function setClients(mongo, redis, metrics) {
+  mongoDbClient = mongo;
+  redisClient = redis;
+  metricsClient = metrics;
+  dataHandlers = dataHandlersModule({
+    defaultHandlers: defaultDataHandlersModule(mongoDbClient)
   });
+  syncStorage = storageModule(mongoDbClient);
+  // TODO: follow same pattern as other modules for this hashProviderModule
+  hashProvider = hashProviderModule;
+  syncLock = syncLockModule(mongoDbClient, 'fhsync_locks');
+  interceptors = interceptorsModule();
 }
-
 
 /**
  * Starts all sync queues, workers & the sync scheduler.
@@ -273,8 +239,8 @@ function callHandler(handlerName, args) {
 module.exports = {
   sync: sync,
   syncRecords: syncRecords,
+  setClients: setClients,
   api: {
-    connect: connect,
     init: init,
     start: start,
     stop: stop,
